处理组网和控制指令协议

+++ src/main/java/com/boe/esl/vo/LabelVO.java
@@ -16,7 +16,7 @@ public class LabelVO implements BaseVO {
 	private String gatewayMac;
 	private String code;
 	private String type;
-	private Short status;
+	private Integer status;
 	private String statusText;
 	private String power;
